Code refactoring to split the code
